requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
google-adk
langchain-community==0.0.8
tavily-python==0.3.0
sympy==1.12
matplotlib==3.8.2
numpy==1.25.2
jinja2
